+++ packages/types/src/client.ts
@@ -70,4 +70,7 @@ export interface Client<O extends Options = Options> {
 
   /** Returns an array of installed integrations on the client. */
   getIntegration<T extends Integration>(integartion: IntegrationClass<T>): T | null;
+
+  /** This is an internal function to setup all integrations that should run on the client */
+  setupIntegrations(): void;
 }